Update code to version 30.0.0
OBS 29.1 breaks API from 29.0:

* update to new API: add new mandatory callback, and add API
* fix new warnings

Tested on: OBS 29.1.2

+++ src/nudgis-config.hpp
@@ -3,6 +3,7 @@
 
 #include <string>
 #include <vector>
+#include <cstdint>
 
 #define DEF_FILE_NAME "nudgis_config.json"
 #define DEF_CHANNEL "Personal channel"
@@ -33,7 +34,7 @@ public:
 	std::string upload_channel;
 	const AutoState *auto_delete_uploaded_file;
 	const AutoState *publish_recording_automatically;
-	uint64_t upload_chunk_size;
+	long long int upload_chunk_size;
 	NudgisConfig();
 	void load(const char *filename = DEF_FILE_NAME);
 	void save(const char *filename = DEF_FILE_NAME);
